requirements.txt
selenium==4.12.0
webdriver-manager==4.0.0
undetected-chromedriver==3.4.6
python-dotenv==1.0.0